Readability & empty OAuthToken prevention
Was passing an empty string for the oAuthToken if one hadn't been
obtained instead of simply not executing the HTTP request. Now return a
Cmd.none in these instances.

Added some convenience functions to avoid having to map using the record
constructor for playlist as usually we only want to modify the isNew
field